+++ lab4/kernel/io.h
@@ -5,7 +5,7 @@ char read_c();
 void read_s(char *s);
 void print_char(const char);
 void print_string(const char *);
-void print_h(const int x);
+void print_h(const unsigned long long x);
 void print_d(const int x);
 
 #endif
